removed mvp e2e folder, moved still needed tests to phase-3, still need to update the tests
