+++ src/applications/check-in/TODO:.md
@@ -1,0 +1,4 @@
+TODO: 
+
+- e2e 
+  - update info feature flip
