+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/app-enabled/disabled.cypress.spec.js
@@ -7,9 +7,6 @@ describe('Check In Experience -- ', () => {
       '/v0/feature_toggles*',
       generateFeatureToggles({
         checkInExperienceEnabled: false,
-        checkInExperienceLowAuthenticationEnabled: false,
-        checkInExperienceMultipleAppointmentSupport: false,
-        checkInExperienceUpdateInformationPageEnabled: false,
       }),
     );
   });
