+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/app-enabled/enabled.cypress.spec.js
@@ -1,6 +1,6 @@
 import { generateFeatureToggles } from '../../../../api/local-mock-api/mocks/feature.toggles';
-import mockCheckIn from '../../../../api/local-mock-api/mocks/v0/check.in.responses';
-import mockValidate from '../../../../api/local-mock-api/mocks/v0/validate.responses';
+import mockSession from '../../../../api/local-mock-api/mocks/v2/sessions.responses';
+import Timeouts from 'platform/testing/e2e/timeouts';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
@@ -8,16 +8,14 @@ describe('Check In Experience -- ', () => {
       'GET',
       '/v0/feature_toggles*',
       generateFeatureToggles({
-        checkInExperienceLowAuthenticationEnabled: false,
-        checkInExperienceMultipleAppointmentSupport: false,
+        checkInExperienceMultipleAppointmentSupport: true,
         checkInExperienceUpdateInformationPageEnabled: false,
       }),
     );
-    cy.intercept('GET', '/check_in/v0/patient_check_ins/*', req => {
-      req.reply(mockValidate.createMockSuccessResponse({}));
-    });
-    cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(mockCheckIn.createMockSuccessResponse({}));
+    cy.intercept('GET', '/check_in/v2/sessions/*', req => {
+      req.reply(
+        mockSession.createMockSuccessResponse('some-token', 'read.basic'),
+      );
     });
   });
   afterEach(() => {
@@ -29,6 +27,8 @@ describe('Check In Experience -- ', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
-    cy.get('h1').contains('Your appointment');
+    cy.get('h1', { timeout: Timeouts.slow })
+      .should('be.visible')
+      .and('have.text', 'Check in at VA');
   });
 });
