+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/errors/check-in-failed.cypress.spec.js
@@ -1,20 +1,16 @@
 import { generateFeatureToggles } from '../../../../api/local-mock-api/mocks/feature.toggles';
-import mockCheckIn from '../../../../api/local-mock-api/mocks/v0/check.in.responses';
-import mockValidate from '../../../../api/local-mock-api/mocks/v0/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/v2/check.in.responses';
+import mockValidate from '../../../../api/local-mock-api/mocks/v2/sessions.responses';
+import Timeouts from 'platform/testing/e2e/timeouts';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply({
-        statusCode: 200,
-        body: mockValidate.createMockFailedResponse({}),
-        delay: 10, // milliseconds
-      });
+      req.reply(mockValidate.createMockSuccessResponse({}));
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(mockCheckIn.createMockSuccessResponse({}));
+      req.reply(mockCheckIn.createMockFailedResponse({}));
     });
-
     cy.intercept(
       'GET',
       '/v0/feature_toggles*',
@@ -30,10 +26,14 @@ describe('Check In Experience -- ', () => {
       window.sessionStorage.clear();
     });
   });
-  it('C5738 - Token is not valid', () => {
+  it('C5722 - Check in failed with a 200 and error message in the body', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
-    cy.get('h1').contains('We couldn’t check you in');
+    cy.get('h1').contains('Your appointment');
+    cy.get('.usa-button').click();
+    cy.get('h1', { timeout: Timeouts.slow })
+      .should('be.visible')
+      .and('contain', 'We couldn’t check you in');
   });
 });
