+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/errors/no.token.provided.cypress.spec.js
@@ -1,16 +1,20 @@
 import { generateFeatureToggles } from '../../../../api/local-mock-api/mocks/feature.toggles';
-
-import mockCheckIn from '../../../../api/local-mock-api/mocks/v0/check.in.responses';
-import mockValidate from '../../../../api/local-mock-api/mocks/v0/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/v2/check.in.responses';
+import mockValidate from '../../../../api/local-mock-api/mocks/v2/sessions.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply(mockValidate.createMockSuccessResponse({}));
+      req.reply({
+        statusCode: 200,
+        body: mockValidate.createMockFailedResponse({}),
+        delay: 10, // milliseconds
+      });
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
       req.reply(mockCheckIn.createMockSuccessResponse({}));
     });
+
     cy.intercept(
       'GET',
       '/v0/feature_toggles*',
@@ -26,16 +30,9 @@ describe('Check In Experience -- ', () => {
       window.sessionStorage.clear();
     });
   });
-  it('C5746 - Happy path', () => {
-    const featureRoute =
-      '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
+  it('C5726 - No token provided', () => {
+    const featureRoute = '/health-care/appointment-check-in/';
     cy.visit(featureRoute);
-    cy.get('h1').contains('Your appointment');
-    cy.injectAxe();
-    cy.axeCheck();
-    cy.get('.usa-button').click();
-    cy.get('va-alert > h1').contains('checked in');
-    cy.injectAxe();
-    cy.axeCheck();
+    cy.get('h1').contains('We couldn’t check you in');
   });
 });
