+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/errors/server.404.on.check-in.cypress.spec.js
@@ -1,20 +1,15 @@
 import { generateFeatureToggles } from '../../../../api/local-mock-api/mocks/feature.toggles';
-import mockCheckIn from '../../../../api/local-mock-api/mocks/v0/check.in.responses';
-import mockValidate from '../../../../api/local-mock-api/mocks/v0/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/v2/check.in.responses';
+import mockValidate from '../../../../api/local-mock-api/mocks/v2/sessions.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply({
-        statusCode: 200,
-        body: mockValidate.createMockFailedResponse({}),
-        delay: 10, // milliseconds
-      });
+      req.reply(mockValidate.createMockSuccessResponse({}));
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(mockCheckIn.createMockSuccessResponse({}));
+      req.reply(404, mockCheckIn.createMockFailedResponse({}));
     });
-
     cy.intercept(
       'GET',
       '/v0/feature_toggles*',
@@ -24,16 +19,22 @@ describe('Check In Experience -- ', () => {
         checkInExperienceUpdateInformationPageEnabled: false,
       }),
     );
+    cy.window().then(window => {
+      window.sessionStorage.clear();
+    });
   });
   afterEach(() => {
     cy.window().then(window => {
       window.sessionStorage.clear();
     });
   });
-  it('C5724 - Token is not valid', () => {
+  it('C5730 - Check in - 404 api error', () => {
     const featureRoute =
-      '/health-care/appointment-check-in/?id=MALFORMED_TOKEN';
+      '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
+    cy.get('h1').contains('Your appointment');
+    cy.get('.usa-button').click();
+    cy.url().should('match', /error/);
     cy.get('h1').contains('We couldn’t check you in');
   });
 });
