+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/errors/server.404.on.validate.cypress.spec.js
@@ -1,15 +1,14 @@
 import { generateFeatureToggles } from '../../../../api/local-mock-api/mocks/feature.toggles';
-
-import mockCheckIn from '../../../../api/local-mock-api/mocks/v0/check.in.responses';
-import mockValidate from '../../../../api/local-mock-api/mocks/v0/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/v2/check.in.responses';
+import mockValidate from '../../../../api/local-mock-api/mocks/v2/sessions.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply(mockValidate.createMockSuccessResponse({}));
+      req.reply(404, mockValidate.createMockFailedResponse({}));
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(mockCheckIn.createMockSuccessResponse({}));
+      req.reply(mockCheckIn.createMockFailedResponse({}));
     });
     cy.intercept(
       'GET',
@@ -17,25 +16,23 @@ describe('Check In Experience -- ', () => {
       generateFeatureToggles({
         checkInExperienceLowAuthenticationEnabled: false,
         checkInExperienceMultipleAppointmentSupport: false,
-        checkInExperienceUpdateInformationPageEnabled: true,
+        checkInExperienceUpdateInformationPageEnabled: false,
       }),
     );
+    cy.window().then(window => {
+      window.sessionStorage.clear();
+    });
   });
   afterEach(() => {
     cy.window().then(window => {
       window.sessionStorage.clear();
     });
   });
-  it('C5684 - Back button from check in page', () => {
+  it('C5732 - Validate - 404 error', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
-    cy.get('legend > h1').contains('information');
-
-    cy.get('[data-testid="yes-button"]').click();
-    cy.get('h1').contains('staff member');
-
-    cy.get('[data-testid="back-button"]').click();
-    cy.get('legend > h1').contains('information');
+    cy.url().should('match', /error/);
+    cy.get('h1').contains('We couldn’t check you in');
   });
 });
