+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/errors/server.500.on.validate.cypress.spec.js
@@ -1,14 +1,14 @@
 import { generateFeatureToggles } from '../../../../api/local-mock-api/mocks/feature.toggles';
-import mockCheckIn from '../../../../api/local-mock-api/mocks/v0/check.in.responses';
-import mockValidate from '../../../../api/local-mock-api/mocks/v0/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/v2/check.in.responses';
+import mockValidate from '../../../../api/local-mock-api/mocks/v2/sessions.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply(mockValidate.createMockSuccessResponse({}));
+      req.reply(500, mockValidate.createMockFailedResponse({}));
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(500, mockCheckIn.createMockFailedResponse({}));
+      req.reply(mockCheckIn.createMockFailedResponse({}));
     });
     cy.intercept(
       'GET',
@@ -28,12 +28,10 @@ describe('Check In Experience -- ', () => {
       window.sessionStorage.clear();
     });
   });
-  it('C5734 - Check in - 500 api error', () => {
+  it('C5736 - Validate - 500 api error', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
-    cy.get('h1').contains('Your appointment');
-    cy.get('.usa-button').click();
     cy.url().should('match', /error/);
     cy.get('h1').contains('We couldn’t check you in');
   });
