+++ src/applications/check-in/tests/e2e/phase-3-multiple-appointments/errors/token.is.not.valid.cypress.spec.js
@@ -1,15 +1,20 @@
 import { generateFeatureToggles } from '../../../../api/local-mock-api/mocks/feature.toggles';
-import mockCheckIn from '../../../../api/local-mock-api/mocks/v0/check.in.responses';
-import mockValidate from '../../../../api/local-mock-api/mocks/v0/validate.responses';
+import mockCheckIn from '../../../../api/local-mock-api/mocks/v2/check.in.responses';
+import mockValidate from '../../../../api/local-mock-api/mocks/v2/sessions.responses';
 
 describe('Check In Experience -- ', () => {
   beforeEach(function() {
     cy.intercept('GET', '/check_in/v0/patient_check_ins//*', req => {
-      req.reply(500, mockValidate.createMockFailedResponse({}));
+      req.reply({
+        statusCode: 200,
+        body: mockValidate.createMockFailedResponse({}),
+        delay: 10, // milliseconds
+      });
     });
     cy.intercept('POST', '/check_in/v0/patient_check_ins/', req => {
-      req.reply(mockCheckIn.createMockFailedResponse({}));
+      req.reply(mockCheckIn.createMockSuccessResponse({}));
     });
+
     cy.intercept(
       'GET',
       '/v0/feature_toggles*',
@@ -19,20 +24,16 @@ describe('Check In Experience -- ', () => {
         checkInExperienceUpdateInformationPageEnabled: false,
       }),
     );
-    cy.window().then(window => {
-      window.sessionStorage.clear();
-    });
   });
   afterEach(() => {
     cy.window().then(window => {
       window.sessionStorage.clear();
     });
   });
-  it('C5736 - Validate - 500 api error', () => {
+  it('C5738 - Token is not valid', () => {
     const featureRoute =
       '/health-care/appointment-check-in/?id=46bebc0a-b99c-464f-a5c5-560bc9eae287';
     cy.visit(featureRoute);
-    cy.url().should('match', /error/);
     cy.get('h1').contains('We couldn’t check you in');
   });
 });
